Kill HeaderResponse; Add a parameter to .header()
The HeaderResponse class was an attempt to copy the perl5 `wantarray` magic.
This doesn't really make sense, and can cause problems when calling modules do not
expect a list to be returned. Instead, we return a string or an array based on
the :multi parameter to the .header call.

Version bump to 2.0, since this is an incompatible change. I don't expect this to
break most code using this module however, since most email headers aren't repeated.